+++ uhabits-android/src/main/java/org/isoron/uhabits/receivers/StartupReceiver.kt
@@ -32,5 +32,5 @@ fun BroadcastReceiver.whenHabitsReady(context: Context, intent: Intent, action: 
             }
         }
     }
-    if (finished) subscription.close()
+    if (finished) subscription?.close()
 }